--- data/googleimages/deerdownload.py
@@ -1,7 +1,0 @@
-from google_images_download import google_images_download   #importing the library
-
-response = google_images_download.googleimagesdownload()   #class instantiation
-
-arguments = {"keywords":"deer","limit":200,"print_urls":True, "format":"jpg", "chromedriver":"/Applications/Google\ Chrome.app/Contents/MacOS/Google\ Chrome"}   #creating list of arguments
-paths = response.download(arguments)   #passing the arguments to the function
-print(paths)   #printing absolute paths of the downloaded images
