+++ subtract_background/readME.md
@@ -10,7 +10,15 @@ Different possible algorithms:
         - outputs a grayscale image of thresholding
         - detects way too many differences (i.e. pixel by pixel comparison, unclean)
 
-    - opencv\_picdiff.py
+    - opencv\_imgsubtraction.py
         - outputs a "subtracted" image
         - still detects ALL the differences (because literal pixel subtraction)
 
+    - bg\_subtract\_video.py
+        - processes a video and does background subtraction
+        - based on movement
+        - may not work on little movement? depends on thresholding
+
+    - background_subtraction.py
+        - this program uses background subtraction methods from OpenCV. 
+        - you can process both videos and images
